+++ modules/crypto/include/ifm3d/crypto.h
@@ -1,0 +1,12 @@
+// -*- c++ -*-
+/*
+ * Copyright 2025-present ifm electronic, gmbh
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#ifndef IFM3D_CRYPTO_H
+#define IFM3D_CRYPTO_H
+
+#include <ifm3d/crypto/crypto.h>
+
+#endif // IFM3D_CRYPTO_H